backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
